Created new script track.py and moved Track class to new script. Imported track class into monitor.py
